+++ channel_test.go
@@ -1,6 +1,7 @@
 package lo
 
 import (
+	"math/rand"
 	"testing"
 	"time"
 
@@ -111,7 +112,21 @@ func TestDispatchingStrategyRoundRobin(t *testing.T) {
 }
 
 func TestDispatchingStrategyRandom(t *testing.T) {
-	// @TODO
+	testWithTimeout(t, 10*time.Millisecond)
+	is := assert.New(t)
+
+	// with this seed, the order of random channels are: 1 - 0
+	rand.Seed(14)
+
+	children := createChannels[int](2, 2)
+	rochildren := channelsToReadOnly(children)
+	defer closeChannels(children)
+
+	for i := 0; i < 2; i++ {
+		children[1] <- i
+	}
+
+	is.Equal(0, DispatchingStrategyRandom(42, 0, rochildren))
 }
 
 func TestDispatchingStrategyWeightedRandom(t *testing.T) {
@@ -242,16 +257,16 @@ func TestGenerate(t *testing.T) {
 	is.Equal(i, 4)
 }
 
-func TestBatch(t *testing.T) {
+func TestBuffer(t *testing.T) {
 	t.Parallel()
 	testWithTimeout(t, 10*time.Millisecond)
 	is := assert.New(t)
 
 	ch := SliceToChannel(2, []int{1, 2, 3})
 
-	items1, length1, _, ok1 := Batch(ch, 2)
-	items2, length2, _, ok2 := Batch(ch, 2)
-	items3, length3, _, ok3 := Batch(ch, 2)
+	items1, length1, _, ok1 := Buffer(ch, 2)
+	items2, length2, _, ok2 := Buffer(ch, 2)
+	items3, length3, _, ok3 := Buffer(ch, 2)
 
 	is.Equal([]int{1, 2}, items1)
 	is.Equal(2, length1)
@@ -264,7 +279,7 @@ func TestBatch(t *testing.T) {
 	is.False(ok3)
 }
 
-func TestBatchWithTimeout(t *testing.T) {
+func TestBufferWithTimeout(t *testing.T) {
 	t.Parallel()
 	testWithTimeout(t, 200*time.Millisecond)
 	is := assert.New(t)
@@ -277,28 +292,99 @@ func TestBatchWithTimeout(t *testing.T) {
 	}
 	ch := Generator(0, generator)
 
-	items1, length1, _, ok1 := BatchWithTimeout(ch, 20, 15*time.Millisecond)
+	items1, length1, _, ok1 := BufferWithTimeout(ch, 20, 15*time.Millisecond)
 	is.Equal([]int{0, 1}, items1)
 	is.Equal(2, length1)
 	is.True(ok1)
 
-	items2, length2, _, ok2 := BatchWithTimeout(ch, 20, 2*time.Millisecond)
+	items2, length2, _, ok2 := BufferWithTimeout(ch, 20, 2*time.Millisecond)
 	is.Equal([]int{}, items2)
 	is.Equal(0, length2)
 	is.True(ok2)
 
-	items3, length3, _, ok3 := BatchWithTimeout(ch, 1, 30*time.Millisecond)
+	items3, length3, _, ok3 := BufferWithTimeout(ch, 1, 30*time.Millisecond)
 	is.Equal([]int{2}, items3)
 	is.Equal(1, length3)
 	is.True(ok3)
 
-	items4, length4, _, ok4 := BatchWithTimeout(ch, 2, 25*time.Millisecond)
+	items4, length4, _, ok4 := BufferWithTimeout(ch, 2, 25*time.Millisecond)
 	is.Equal([]int{3, 4}, items4)
 	is.Equal(2, length4)
 	is.True(ok4)
 
-	items5, length5, _, ok5 := BatchWithTimeout(ch, 3, 25*time.Millisecond)
+	items5, length5, _, ok5 := BufferWithTimeout(ch, 3, 25*time.Millisecond)
 	is.Equal([]int{}, items5)
 	is.Equal(0, length5)
 	is.False(ok5)
+}
+
+func TestFanIn(t *testing.T) {
+	t.Parallel()
+	testWithTimeout(t, 100*time.Millisecond)
+	is := assert.New(t)
+
+	upstreams := createChannels[int](3, 10)
+	roupstreams := channelsToReadOnly(upstreams)
+	for i := range roupstreams {
+		go func(i int) {
+			upstreams[i] <- 1
+			upstreams[i] <- 1
+			close(upstreams[i])
+		}(i)
+	}
+	out := FanIn(10, roupstreams...)
+	time.Sleep(10 * time.Millisecond)
+
+	// check input channels
+	is.Equal(0, len(roupstreams[0]))
+	is.Equal(0, len(roupstreams[1]))
+	is.Equal(0, len(roupstreams[2]))
+
+	// check channels allocation
+	is.Equal(6, len(out))
+	is.Equal(10, cap(out))
+
+	// check channels content
+	for i := 0; i < 6; i++ {
+		msg0, ok0 := <-out
+		is.Equal(true, ok0)
+		is.Equal(1, msg0)
+	}
+
+	// check it is closed
+	time.Sleep(10 * time.Millisecond)
+	msg0, ok0 := <-out
+	is.Equal(false, ok0)
+	is.Equal(0, msg0)
+}
+
+func TestFanOut(t *testing.T) {
+	t.Parallel()
+	testWithTimeout(t, 100*time.Millisecond)
+	is := assert.New(t)
+
+	upstream := SliceToChannel(10, []int{0, 1, 2, 3, 4, 5})
+	rodownstreams := FanOut(3, 10, upstream)
+
+	time.Sleep(10 * time.Millisecond)
+
+	// check output channels
+	is.Equal(3, len(rodownstreams))
+
+	// check channels allocation
+	for i := range rodownstreams {
+		is.Equal(6, len(rodownstreams[i]))
+		is.Equal(10, cap(rodownstreams[i]))
+		is.Equal([]int{0, 1, 2, 3, 4, 5}, ChannelToSlice(rodownstreams[i]))
+	}
+
+	// check it is closed
+	time.Sleep(10 * time.Millisecond)
+
+	// check channels allocation
+	for i := range rodownstreams {
+		msg, ok := <-rodownstreams[i]
+		is.Equal(false, ok)
+		is.Equal(0, msg)
+	}
 }