+++ string_test.go
@@ -2,10 +2,34 @@ package lo
 
 import (
 	"math"
+	"math/rand"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
+
+func TestRandomString(t *testing.T) {
+	t.Parallel()
+	is := assert.New(t)
+
+	rand.Seed(time.Now().UnixNano())
+
+	str1 := RandomString(100, LowerCaseLettersCharset)
+	is.Equal(100, RuneLength(str1))
+	is.Subset(LowerCaseLettersCharset, []rune(str1))
+
+	str2 := RandomString(100, LowerCaseLettersCharset)
+	is.NotEqual(str1, str2)
+
+	noneUtf8Charset := []rune("明1好休2林森")
+	str3 := RandomString(100, noneUtf8Charset)
+	is.Equal(100, RuneLength(str3))
+	is.Subset(noneUtf8Charset, []rune(str3))
+
+	is.PanicsWithValue("lo.RandomString: Charset parameter must not be empty", func() { RandomString(100, []rune{}) })
+	is.PanicsWithValue("lo.RandomString: Size parameter must be greater than 0", func() { RandomString(0, LowerCaseLettersCharset) })
+}
 
 func TestChunkString(t *testing.T) {
 	t.Parallel()